Cross-validation for NMF

Find an "optimal" rank for a Non-Negative Matrix Factorization using cross-validation. Returns a data.frame with class nmfCrossValidate. Plot results using the plot S3 method.

crossValidate(data, k, method = "impute", reps = 5, n = 0.05, ...)

# S3 method for nmfCrossValidate
plot(x, ...)

Arguments

data
dense or sparse matrix of features in rows and samples in columns. Prefer matrix or Matrix::dgCMatrix, respectively

k
array of factorization ranks to test

method
algorithm to use for cross-validation: "predict" = bi-cross-validation on equally sized feature/sample subsets, "impute" = MSE of missing value imputation, "robust" = mean cost of bipartite matching between models trained on equally sized non-overlapping sample sets.

reps
number of independent replicates to run

n
for method = "impute" and "perturb", fraction of values to handle as missing

...
parameters to RcppML::nmf, not including data or k

x
nmfCrossValidate object, the result of crossValidate

Value

data.frame with columns rep, k, and value, where value depends on the method selected (i.e. MSE, cost of bipartite matching)

Details

Two algorithms are available for cross-validation and can be selected using the method parameter:

Bi-cross-validation: Data is split into an equally-sized 2x2 grid (A1, A2, B1, and B2). Data is trained on A1 and projected to A2 to B2 where the mean squared error of the model on B2 is measured.

Cost of bipartite matching between independent replicates. Samples are partitioned into two sets and independently factorized, then correlation between factors in both w models is computed, factors are matched using bipartite matching on a cosine similarity cost matrix, and the mean cost of matched factors is returned.

See also