Project an NMF model onto new samples

Given an NMF model in the form \(A = wdh\), predict.nmf projects w onto A to solve for h.

# S4 method for nmf
predict(object, data, L1 = 0, L2 = 0, nonneg = TRUE, mask = NULL, ...)

predict.nmf(w, data, L1 = 0, L2 = 0, nonneg = TRUE, mask = NULL, ...)

Arguments

object
fitted model, class nmf, generally the result of calling nmf, with models of equal dimensions as data

data
dense or sparse matrix of features in rows and samples in columns. Prefer matrix or Matrix::dgCMatrix, respectively

L1
a single LASSO penalty in the range (0, 1]

L2
a single Ridge penalty greater than zero

nonneg
enforce non-negativity

mask
dense or sparse matrix of values in data to handle as missing. Prefer Matrix::ngCMatrix. Alternatively, specify "zeros" or "NA" to mask either all zeros or NA values.

...
arguments passed to or from other methods

w
matrix of features (rows) by factors (columns), corresponding to rows in data

Value

object of class nmf

Details

TO DO: add documentation For the alternating least squares matrix factorization update problem \(A = wh\), the updates (or projection) of \(h\) is given by the equation: $$w^Twh = wA_j$$ which is in the form \(ax = b\) where \(a = w^Tw\) \(x = h\) and \(b = wA_j\) for all columns \(j\) in \(A\).

Any L1 penalty is subtracted from \(b\) and should generally be scaled to max(b), where \(b = WA_j\) for all columns \(j\) in \(A\). An easy way to properly scale an L1 penalty is to normalize all columns in \(w\) to sum to the same value (e.g. 1). No scaling is applied in this function. Such scaling guarantees that L1 = 1 gives a completely sparse solution.

There are specializations for dense and sparse input matrices, symmetric input matrices, and for rank-1 and rank-2 projections. See documentation for nmf for theoretical details and guidance.

References

DeBruine, ZJ, Melcher, K, and Triche, TJ. (2021). "High-performance non-negative matrix factorization for large single-cell data." BioRXiv.

See also

Author

Zach DeBruine

Examples