Project a factor model with least squares

Solves the equation \(A = wh\) for \(h\) given \(w\) and \(A\)

project(model, newdata, L1 = 0, nonneg = TRUE, mask_zeros = FALSE)

Arguments

model
\(w\), matrix of features (rows) by factors (columns) containing the linear model to project

newdata
\(A\), data containing features (rows) by samples (columns) containing the new samples onto which to project the model

L1
L1/LASSO penalty to be applied. No scaling is performed. See details.

nonneg
enforce non-negativity

mask_zeros
handle zeros as missing values

Value

matrix \(h\)

Details

For the alternating least squares matrix factorization update problem \(A = wh\), the updates (or projection) of \(h\) is given by the equation: $$w^Twh = wA_j$$ which is in the form \(ax = b\) where \(a = w^Tw\) \(x = h\) and \(b = wA_j\) for all columns \(j\) in \(A\).

Least squares projections in factorizations of rank-3 and greater are parallelized using the number of threads set by setRcppMLthreads.

Any L1 penalty is subtracted from \(b\) and should generally be scaled to max(b), where \(b = WA_j\) for all columns \(j\) in \(A\). An easy way to properly scale an L1 penalty is to normalize all columns in \(w\) to sum to the same value (e.g. 1). No scaling is applied in this function. Such scaling guarantees that L1 = 1 gives a completely sparse solution.

There are specializations for dense and sparse input matrices, symmetric input matrices, and for rank-1 and rank-2 projections. See documentation for nmf for theoretical details and guidance.

References

DeBruine, ZJ, Melcher, K, and Triche, TJ. (2021). "High-performance non-negative matrix factorization for large single-cell data." BioRXiv.

See also

Author

Zach DeBruine

Examples

if (FALSE) {
library(Matrix)
w <- matrix(runif(1000 * 10), 1000, 10)
h_true <- matrix(runif(10 * 100), 10, 100)
# A is the crossproduct of "w" and "h" with 10% signal dropout
A <- (w %*% h_true) * (rsparsematrix(1000, 100, 0.9) > 0)
h <- project(A, w)
cor(as.vector(h_true), as.vector(h))

# alternating projections refine solution (like NMF)
mse_bad <- mse(A, w, rep(1, ncol(w)), h) # mse before alternating updates
h <- project(A, w = w)
w <- project(A, h = h)
h <- project(A, w)
w <- project(A, h = h)
h <- project(A, w)
w <- t(project(A, h = h))
mse_better <- mse(A, w, rep(1, ncol(w)), h) # mse after alternating updates
mse_better < mse_bad

# two ways to solve for "w" that give the same solution
w <- project(A, h = h)
w2 <- project(t(A), w = t(h))
all.equal(w, w2)
}