Summarize NMF factors

summary method for class "nmf". Describes metadata representation in NMF factors. Returns object of class nmfSummary. Plot results using plot.

# S4 method for nmf
summary(object, group_by, stat = "sum", ...)

# S3 method for nmfSummary
plot(x, ...)

Arguments

object
an object of class "nmf", usually, a result of a call to nmf

group_by
a discrete factor giving groupings for samples or features. Must be of the same length as number of samples in object$h or number of features in object$w.

stat
either sum (sum of factor weights falling within each group), or mean (mean factor weight falling within each group).

...
arguments passed to or from other methods

x
nmfSummary object, the result of calling summary on an nmf object

Value

data.frame with columns group, factor, and stat